requirements.txt
python-magic-bin == 0.4.14
requests